+++ apps/dav/lib/Connector/Sabre/QuotaPlugin.php
@@ -29,6 +29,7 @@ use OCP\Files\StorageNotAvailableException;
 use Sabre\DAV\Exception\InsufficientStorage;
 use Sabre\DAV\Exception\ServiceUnavailable;
 use Sabre\DAV\INode;
+use Sabre\HTTP\RequestInterface;
 
 /**
  * This plugin check user quota and deny creating files when they exceeds the quota.
@@ -75,6 +76,7 @@ class QuotaPlugin extends \Sabre\DAV\ServerPlugin {
 		$server->on('beforeWriteContent', [$this, 'handleBeforeWriteContent'], 10);
 		$server->on('beforeCreateFile', [$this, 'handleBeforeCreateFile'], 10);
 		$server->on('beforeMove', [$this, 'handleBeforeMove'], 10);
+		$server->on('beforeCopy', [$this, 'handleBeforeCopy'], 10);
 	}
 
 	/**
@@ -90,8 +92,12 @@ class QuotaPlugin extends \Sabre\DAV\ServerPlugin {
 		$sourceNode = $this->server->tree->getNodeForPath($source);
 
 		// get target node for proper path conversion
+		$destinationSize = 0;
 		if ($this->server->tree->nodeExists($destination)) {
 			$targetNode = $this->server->tree->getNodeForPath($destination);
+			if ($targetNode instanceof Node) {
+				$destinationSize = $targetNode->getSize() ?? 0;
+			}
 		} else {
 			$dirname = \dirname($destination);
 			$dirname = $dirname === '.' ? '/' : $dirname;
@@ -112,7 +118,31 @@ class QuotaPlugin extends \Sabre\DAV\ServerPlugin {
 			}
 		}
 		$path = $targetNode->getPath();
-		return $this->checkQuota($path, $sourceNode->getSize());
+		return $this->checkQuota($path, $sourceNode->getSize(), $destinationSize);
+	}
+
+	public function handleBeforeCopy($source, $destination) {
+		$sourceNode = $this->server->tree->getNodeForPath($source); // might throw a NotFound exception
+
+		$sourceSize = null;
+		if ($sourceNode instanceof Node) {
+			$sourceSize = $sourceNode->getSize();
+		}
+
+		$destinationSize = 0;
+		if ($this->server->tree->nodeExists($destination)) {
+			$targetNode = $this->server->tree->getNodeForPath($destination);
+			if ($targetNode instanceof Node) {
+				$destinationSize = $targetNode->getSize() ?? 0;
+			}
+		} else {
+			$dirname = \dirname($destination);
+			$dirname = $dirname === '.' ? '/' : $dirname;
+			$targetNode = $this->server->tree->getNodeForPath($dirname);
+		}
+
+		$path = $targetNode->getPath();
+		return $this->checkQuota($path, $sourceSize, $destinationSize);
 	}
 
 	/**
@@ -127,7 +157,9 @@ class QuotaPlugin extends \Sabre\DAV\ServerPlugin {
 		if (!$node instanceof Node) {
 			return;
 		}
-		return $this->checkQuota($node->getPath());
+
+		$nodeSize = $node->getSize() ?? 0;  // it might return null, so assume 0 in that case.
+		return $this->checkQuota($node->getPath(), null, $nodeSize);
 	}
 
 	/**
@@ -150,10 +182,11 @@ class QuotaPlugin extends \Sabre\DAV\ServerPlugin {
 	 *
 	 * @param string $path path of the user's home
 	 * @param int $length size to check whether it fits
+	 * @param int $extraSpace additional space granted, usually used when overwriting files
 	 * @throws InsufficientStorage
 	 * @return bool
 	 */
-	public function checkQuota($path, $length = null) {
+	public function checkQuota($path, $length = null, $extraSpace = 0) {
 		if ($length === null) {
 			$length = $this->getLength();
 		}
@@ -173,7 +206,7 @@ class QuotaPlugin extends \Sabre\DAV\ServerPlugin {
 				$path = \rtrim($parentPath, '/') . '/' . $info['name'];
 			}
 			$freeSpace = $this->getFreeSpace($path);
-			if ($freeSpace !== FileInfo::SPACE_UNKNOWN && $freeSpace !== FileInfo::SPACE_UNLIMITED && $length > $freeSpace) {
+			if ($freeSpace !== FileInfo::SPACE_UNKNOWN && $freeSpace !== FileInfo::SPACE_UNLIMITED && $length > $freeSpace + $extraSpace) {
 				if (isset($chunkHandler)) {
 					$chunkHandler->cleanup();
 				}